==> 1_EDA.ipynb <==
# !pip install wandb 

## EDA

In this notebook, we downloaded a sample semantic segmentation dataset and use W&B Artifacts and Tables to version and analyse our data

from fastai.vision.all import * # to download and unzip our datasets
import params
import wandb

# /params.py
# WANDB_PROJECT = "mlops-course-001"
# ENTITY = None # set this to team name if working in a team
# BDD_CLASSES = {i:c for i,c in enumerate(['background', 'road', 'traffic light', 'traffic sign', 'person', 'vehicle', 'bicycle'])}
# RAW_DATA_AT = 'bdd_simple_1k'
# PROCESSED_DATA_AT = 'bdd_simple_1k_split'

URL = "https://storage.googleapis.com/wandb_course/bdd_simple_1k.zip"

path = Path(untar_data(URL, force_download=True))

(path/"labels").ls()

import numpy as np

def label_func(fname):
    return (fname.parent.parent/"labels"/f"{fname.stem}_mask.png")

def get_classes_per_image(mask_data, class_labels):
    unique = list(np.unique(mask_data))
    result_dict = {}
    for _class in class_labels.keys():
        result_dict[class_labels[_class]] = int(_class in unique)
    return result_dict

def _create_table(image_files, class_labels):
    "create a table with the dataset "
    labels = [str(class_labels[_lab] for _lab in list(class_labels))]
    table = wandb.Table(columns=["File_Name", "P1","P2" "Images", "Dataset"] + labels)
    
    for i, image_file in progress_bar(enumerate(image_files), total = len(image_files)):
        image =Image.open(image_file)
        mask_data = np.array(Image.open(label_func(image_file)))
        class_in_image = get_classes_per_image(mask_data, class_labels)
        table.add_data(
            image_file.stem,
            image_file.stem.split('-')[0],
            image_file.stem.split('-')[1],
            wandb.Image(image,
                        masks = {
                            "predictions":{
                                "mask_data":mask_data,
                                 "class_labels":class_labels,
                            }
                        }),
            "bdd1k", 
            *[class_in_image[_lab] for _lab in labels]
        )
    return table


to help us process the data and upload it as a Table to W&B

def label_func(fname):
    return (fname.parent.parent/"labels")/f"{fname.stem}_mask.png"

def get_classes_per_image(mask_data, class_labels):
    unique = list(np.unique(mask_data))
    result_dict = {}
    for _class in class_labels.keys():
        result_dict[class_labels[_class]] = int(_class in unique)
    return result_dict

def _create_table(image_files, class_labels):
    "Create a table with the dataset"
    labels = [str(class_labels[_lab]) for _lab in list(class_labels)]
    table = wandb.Table(columns=["File_Name", "Images", "Split"] + labels)
    
    for i, image_file in progress_bar(enumerate(image_files), total=len(image_files)):
        image = Image.open(image_file)
        mask_data = np.array(Image.open(label_func(image_file)))
        class_in_image = get_classes_per_image(mask_data, class_labels)
        table.add_data(
            str(image_file.name),
            wandb.Image(
                    image,
                    masks={
                        "predictions": {
                            "mask_data": mask_data,
                            "class_labels": class_labels,
                        }
                    }
            ),
            "None", # we don't have a dataset split yet
            *[class_in_image[_lab] for _lab in labels]
        )
    
    return table
     

we will start a new W&B  **run** and put everything into a raw Artifacts

run = wandb.init(project = params.WANDB_PROJECT, entity = params.ENTITY, job_type = "upload")

# CREATE AN ARTIFACT
artifact = wandb.Artifact(params.RAW_DATA_AT, type='raw_data' )

# ADD FILE TO ARTIFACT
artifact.add_file(path/'LICENSE.txt', name="LICENSE.txt")

lets add the images and label masks.

# ADD FOLDERS TO ARTIFACT
artifact.add_dir(path/'images', name='images')
artifact.add_dir(path/'labels', name= 'labels')

lets get the file names of images in our 

image_files = get_image_files(path/"images", recurse = False)

table = _create_table(image_files, params.BDD_CLASSES)

# ADD TABLE TO ARTIFACT
artifact.add(table, "eda_table")

# LOG ARTIFACT
run.log_artifact(artifact)

# FINISH RUN 
run.finish()
