==> 2_split.ipynb <==
## Data Preparation

In this notebook, we will prepare the data to later train our deep learning model. To do so, 
- we willl start a new W&B  **run** and use our data artifact
- split the data and save the splits into a new W&B  Artifacts
- join information about the split with our EDA Table

import os, warnings
import wandb

import pandas as pd
from fastai.vision.all import *
from sklearn.model_selection import StratifiedGroupKFold
import params
warnings.filterwarnings('ignore')

Use the artifact we prevoiusly saved to W&B (we are storing artifact names and other global parameters in params)

run = wandb.init(project=params.WANDB_PROJECT, entity = params.ENTITY, job_type = "data_split")

raw_data_at = run.use_artifact(f'{params.RAW_DATA_AT}:latest')
path = Path(raw_data_at.download())

path.ls()

To split data between training , testing and validation , we need file names , groups or some thing like that  and target (here we use our rare class bicycle for stratification). We previously saved these columns to EDA table, so lets retrieve it from the table now. 

orig_eda_table = raw_data_at.get("eda_table")

fnames = orig_eda_table.get_column("File_Name")
groups = [s.split('-')[0] for s in fnames]
y = orig_eda_table.get_column("bicycle")

Now we will split the data into train(80%) , validataion(10%), and test (10%) sets. As we do that, we need to be carefull to :
- avoid leakage : for that reason we are grouping data according to video identifier (we want to make sure our model can generalize o new cars or video frames)
- handle the label imbalance : for that reason we startify data with our target column 

We will use sklearn's  **StratifiedgroupKFold** to split the data into 10 folds and assign 1 fold for test, 1 for validation and the rest for training.

df = pd.DataFrame()
df["File_Name"] = fnames
df['fold'] = -1

cv = StratifiedGroupKFold(n_splits=10)
for i, (train_idxs, test_idxs) in enumerate(cv.split(fnames, y, groups)):
    df.loc[test_idxs, ['fold']] = i

df.head()

df['Stage'] = 'train'
df.loc[df.fold == 0, ['Stage']] = 'test'
df.loc[df.fold == 1, ['Stage']] = 'valid'
del df['fold']
df.Stage.value_counts()

df.to_csv('data_split.csv', index= False)

Now to create a new Artifact and add our data there.

processed_data_at = wandb.Artifact(params.PROCESSED_DATA_AT, type = "split_data")

processed_data_at.add_file("data_split.csv")
processed_data_at.add_dir(path)

Finally, the split info may be relevant for our analysis - rather than uploading images again , we will save the split info to a new table and join it with EDA table we created prevoiusly.

data_split_table = wandb.Table(dataframe= df[["File_Name", "Stage"]])

join_table = wandb.JoinedTable(orig_eda_table, data_split_table, "File_Name")

Lets add it to our artifact, log it and finish our **run**

processed_data_at.add(join_table, "eda_table_data_split")

run.log_artifact(processed_data_at)
run.finish()
